Split into lib and executable
