+++ caas.cabal
@@ -1,32 +1,44 @@
--- Initial caas.cabal generated by cabal init.  For further 
+-- Initial caas.cabal generated by cabal init.  For further
 -- documentation, see http://haskell.org/cabal/users-guide/
 
 name:                caas
 version:             0.1.0.0
--- synopsis:            
--- description:         
+-- synopsis:
+-- description:
 -- license:             MIT
 -- license-file:        LICENSE
 author:              Sebastian Mielke
 maintainer:          githubcontact@sjmielke.com
--- copyright:           
+-- copyright:
 category:            Web
 build-type:          Simple
--- extra-source-files:  
+-- extra-source-files:
 cabal-version:       >=1.10
+
+library
+  exposed-modules:   CAAS.Fibo
+                   , CAAS.Mosaic
+  build-depends:     base >=4.8 && <5,
+                     random >= 1.1,
+                     mtl >= 2.2,
+                     case-insensitive >= 1.2
+  default-language: Haskell2010
+  hs-source-dirs:   lib
+
 
 executable caas
   main-is:             FiboMain.hs
-  -- other-modules:       
-  -- other-extensions:    
-  build-depends:       
-    base >=4.8 && <4.9,
+  -- other-modules:
+  -- other-extensions:
+  build-depends:
+    base >=4.8 && <5,
     wai >= 3.0,
     warp >= 3.0,
     bytestring >= 0.10,
     random >= 1.1,
     mtl >= 2.2,
     case-insensitive >= 1.2,
-    http-types >= 0.8
+    http-types >= 0.8,
+    caas
   hs-source-dirs:      src
   default-language:    Haskell2010
